cleaned, now full select framework
framework, i.e. my select wrapper wants full control and just hands over
control to custom functions temporarily. This is because it hides all
low-level filedescriptor stuff. For this, it needs to own the TcpStreams
and server socket and forked child.

Next idea: get rid of framework stuff, build a library which allow more
control and flexibility. For this, I want to remove the ReadOnce,
AcceptOnce, ... stuff (though it is quite awesome) and make select
probably take a buffer and return an iterator which has already filled
the buffers for you for reading. Accepting new connections may stay
framework style.